backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pillow==10.1.0
python-multipart==0.0.6
numpy==1.26.1
pydantic==2.4.2
tensorflow==2.12.0
keras==2.12.0
